refactor: renamed consumer & producer naming
