+++ RepositoryCompiler/Communication/MessageProducer.cs
@@ -13,7 +13,7 @@ namespace RepositoryCompiler.Communication
     /// 
     /// We need to refactor and make more suitable infrastructure !
     /// </summary>
-    public class RepositoryCompilerMessageProducer
+    public class MessageProducer
     {
         public string NodeName { get; set; }
         public string QueueName { get; set; }
@@ -21,7 +21,7 @@ namespace RepositoryCompiler.Communication
         public IConnection Connection { get; set; }
         public IModel Channel { get; set; }
 
-        public RepositoryCompilerMessageProducer()
+        public MessageProducer()
         {
             ConfigureInitialStates();
             CreateConnection();
